+++ src/com/intellij/idea/plugin/hybris/system/type/psi/reference/result/AttributeResolveResult.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -19,12 +19,10 @@
 package com.intellij.idea.plugin.hybris.system.type.psi.reference.result
 
 import com.intellij.idea.plugin.hybris.system.type.meta.model.TSGlobalMetaItem
-import com.intellij.idea.plugin.hybris.system.type.model.Attribute
 
 class AttributeResolveResult(
     val meta: TSGlobalMetaItem.TSGlobalMetaItemAttribute
 ) : TSResolveResult {
-    private val myDom: Attribute? = meta.retrieveDom()
-    override fun getElement() = myDom?.qualifier?.xmlAttributeValue
-    override fun isValidResult() = (myDom?.isValid ?: false) && element != null
+    override fun getElement() = meta.retrieveDom()?.qualifier?.xmlAttributeValue
+    override fun isValidResult() = element != null && (meta.retrieveDom()?.isValid ?: false)
 }