+++ src/com/intellij/idea/plugin/hybris/system/type/psi/reference/result/MapResolveResult.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -25,8 +25,7 @@ class MapResolveResult(
     val meta: TSMetaMap,
     private val navigateTo: String = MapType.CODE
 ) : TSResolveResult {
-    private val myDom: MapType? = meta.retrieveDom()
-    override fun getElement() = myDom
+    override fun getElement() = meta.retrieveDom()
         ?.let {
             when (navigateTo) {
                 MapType.CODE -> it.code.xmlAttributeValue
@@ -35,5 +34,5 @@ class MapResolveResult(
                 else -> null
             }
         }
-    override fun isValidResult() = (myDom?.isValid ?: false) && element != null
+    override fun isValidResult() = element != null && (meta.retrieveDom()?.isValid ?: false)
 }