+++ src/com/intellij/idea/plugin/hybris/system/type/psi/reference/result/RelationEndResolveResult.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -19,20 +19,14 @@
 package com.intellij.idea.plugin.hybris.system.type.psi.reference.result
 
 import com.intellij.idea.plugin.hybris.system.type.meta.model.TSMetaRelation
-import com.intellij.idea.plugin.hybris.system.type.model.RelationElement
 import com.intellij.psi.PsiElement
 
 class RelationEndResolveResult(
     val meta: TSMetaRelation.TSMetaRelationElement
 ) : TSResolveResult {
-    private val myDom: RelationElement? = meta.retrieveDom()
-    override fun getElement(): PsiElement? = if (meta.isNavigable) myDom
-        ?.qualifier
-        ?.xmlAttributeValue else {
-        myDom
-            ?.type
-            ?.xmlAttributeValue
-    }
+    override fun getElement(): PsiElement? = meta.retrieveDom()
+        ?.let { if (meta.isNavigable) it.qualifier else it.type }
+        ?.xmlAttributeValue
 
-    override fun isValidResult() = (myDom?.isValid ?: false) && element != null
+    override fun isValidResult() = element != null && (meta.retrieveDom()?.isValid ?: false)
 }